Move extensions to its own package
