+++ app/src/main/java/com/topjohnwu/magisk/data/repository/DBConfig.kt
@@ -2,7 +2,7 @@ package com.topjohnwu.magisk.data.repository
 
 import com.topjohnwu.magisk.data.database.SettingsDao
 import com.topjohnwu.magisk.data.database.StringDao
-import com.topjohnwu.magisk.utils.trimEmptyToNull
+import com.topjohnwu.magisk.extensions.trimEmptyToNull
 import io.reactivex.schedulers.Schedulers
 import kotlin.properties.ReadWriteProperty
 import kotlin.reflect.KProperty
