+++ app/src/main/java/com/topjohnwu/magisk/extensions/XString.kt
@@ -1,4 +1,4 @@
-package com.topjohnwu.magisk.utils
+package com.topjohnwu.magisk.extensions
 
 import android.content.res.Resources
