+++ app/src/main/java/com/topjohnwu/magisk/model/entity/internal/DownloadSubject.kt
@@ -4,10 +4,10 @@ import android.content.Context
 import android.os.Parcelable
 import com.topjohnwu.magisk.Config
 import com.topjohnwu.magisk.Info
+import com.topjohnwu.magisk.extensions.cachedFile
+import com.topjohnwu.magisk.extensions.get
 import com.topjohnwu.magisk.model.entity.MagiskJson
 import com.topjohnwu.magisk.model.entity.module.Repo
-import com.topjohnwu.magisk.utils.cachedFile
-import com.topjohnwu.magisk.utils.get
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
 import java.io.File
