+++ app/src/main/java/com/topjohnwu/magisk/model/entity/recycler/HideRvItem.kt
@@ -6,12 +6,12 @@ import com.skoumal.teanity.rxbus.RxBus
 import com.skoumal.teanity.util.DiffObservableList
 import com.skoumal.teanity.util.KObservableField
 import com.topjohnwu.magisk.R
+import com.topjohnwu.magisk.extensions.inject
+import com.topjohnwu.magisk.extensions.toggle
 import com.topjohnwu.magisk.model.entity.HideAppInfo
 import com.topjohnwu.magisk.model.entity.HideTarget
 import com.topjohnwu.magisk.model.entity.state.IndeterminateState
 import com.topjohnwu.magisk.model.events.HideProcessEvent
-import com.topjohnwu.magisk.utils.inject
-import com.topjohnwu.magisk.utils.toggle
 
 class HideRvItem(val item: HideAppInfo, targets: List<HideTarget>) :
     ComparableRvItem<HideRvItem>() {
