+++ app/src/main/java/com/topjohnwu/magisk/model/entity/recycler/ModuleRvItem.kt
@@ -6,10 +6,10 @@ import com.skoumal.teanity.databinding.ComparableRvItem
 import com.skoumal.teanity.extensions.addOnPropertyChangedCallback
 import com.skoumal.teanity.util.KObservableField
 import com.topjohnwu.magisk.R
+import com.topjohnwu.magisk.extensions.get
+import com.topjohnwu.magisk.extensions.toggle
 import com.topjohnwu.magisk.model.entity.module.Module
 import com.topjohnwu.magisk.model.entity.module.Repo
-import com.topjohnwu.magisk.utils.get
-import com.topjohnwu.magisk.utils.toggle
 
 class ModuleRvItem(val item: Module) : ComparableRvItem<ModuleRvItem>() {
