+++ app/src/main/java/com/topjohnwu/magisk/tasks/FlashZip.kt
@@ -4,9 +4,9 @@ import android.net.Uri
 import com.skoumal.teanity.extensions.subscribeK
 import com.topjohnwu.magisk.App
 import com.topjohnwu.magisk.Const
-import com.topjohnwu.magisk.utils.fileName
-import com.topjohnwu.magisk.utils.inject
-import com.topjohnwu.magisk.utils.readUri
+import com.topjohnwu.magisk.extensions.fileName
+import com.topjohnwu.magisk.extensions.inject
+import com.topjohnwu.magisk.extensions.readUri
 import com.topjohnwu.magisk.utils.unzip
 import com.topjohnwu.superuser.Shell
 import io.reactivex.Single
