+++ app/src/main/java/com/topjohnwu/magisk/tasks/InstallerHelper.kt
@@ -2,8 +2,8 @@ package com.topjohnwu.magisk.tasks
 
 import android.content.Context
 import android.net.Uri
-import com.topjohnwu.magisk.utils.get
-import com.topjohnwu.magisk.utils.readUri
+import com.topjohnwu.magisk.extensions.get
+import com.topjohnwu.magisk.extensions.readUri
 import java.io.File
 
 object InstallerHelper {
