+++ app/src/main/java/com/topjohnwu/magisk/ui/home/HomeViewModel.kt
@@ -6,13 +6,13 @@ import com.skoumal.teanity.extensions.subscribeK
 import com.skoumal.teanity.util.KObservableField
 import com.topjohnwu.magisk.*
 import com.topjohnwu.magisk.data.repository.MagiskRepository
+import com.topjohnwu.magisk.extensions.packageName
+import com.topjohnwu.magisk.extensions.res
+import com.topjohnwu.magisk.extensions.toggle
 import com.topjohnwu.magisk.model.events.*
 import com.topjohnwu.magisk.model.observer.Observer
 import com.topjohnwu.magisk.ui.base.MagiskViewModel
 import com.topjohnwu.magisk.utils.ISafetyNetHelper
-import com.topjohnwu.magisk.utils.packageName
-import com.topjohnwu.magisk.utils.res
-import com.topjohnwu.magisk.utils.toggle
 import com.topjohnwu.superuser.Shell
