+++ app/src/main/java/com/topjohnwu/magisk/ui/module/ModuleViewModel.kt
@@ -10,6 +10,8 @@ import com.skoumal.teanity.util.KObservableField
 import com.topjohnwu.magisk.BR
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.data.database.RepoDao
+import com.topjohnwu.magisk.extensions.toSingle
+import com.topjohnwu.magisk.extensions.update
 import com.topjohnwu.magisk.model.entity.module.Module
 import com.topjohnwu.magisk.model.entity.recycler.ModuleRvItem
 import com.topjohnwu.magisk.model.entity.recycler.RepoRvItem
@@ -19,8 +21,6 @@ import com.topjohnwu.magisk.model.events.OpenChangelogEvent
 import com.topjohnwu.magisk.model.events.OpenFilePickerEvent
 import com.topjohnwu.magisk.tasks.RepoUpdater
 import com.topjohnwu.magisk.ui.base.MagiskViewModel
-import com.topjohnwu.magisk.utils.toSingle
-import com.topjohnwu.magisk.utils.update
 import io.reactivex.Single
 import io.reactivex.disposables.Disposable
 import me.tatarka.bindingcollectionadapter2.OnItemBind
