+++ app/src/main/java/com/topjohnwu/magisk/ui/module/ReposFragment.kt
@@ -11,9 +11,9 @@ import com.topjohnwu.magisk.Config
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.FragmentReposBinding
 import com.topjohnwu.magisk.model.download.DownloadService
-import com.topjohnwu.magisk.model.entity.module.Repo
 import com.topjohnwu.magisk.model.entity.internal.Configuration
 import com.topjohnwu.magisk.model.entity.internal.DownloadSubject
+import com.topjohnwu.magisk.model.entity.module.Repo
 import com.topjohnwu.magisk.model.events.InstallModuleEvent
 import com.topjohnwu.magisk.model.events.OpenChangelogEvent
 import com.topjohnwu.magisk.ui.base.MagiskFragment
