+++ app/src/main/java/com/topjohnwu/magisk/utils/DataBindingAdapters.kt
@@ -18,6 +18,7 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.navigation.NavigationView
 import com.skoumal.teanity.extensions.subscribeK
 import com.topjohnwu.magisk.R
+import com.topjohnwu.magisk.extensions.replaceRandomWithSpecial
 import com.topjohnwu.magisk.model.entity.state.IndeterminateState
 import io.reactivex.Observable
 import io.reactivex.disposables.Disposable
