+++ app/src/main/java/com/topjohnwu/magisk/utils/RootUtils.kt
@@ -7,6 +7,8 @@ import android.net.Uri
 import com.topjohnwu.magisk.Const
 import com.topjohnwu.magisk.Info
 import com.topjohnwu.magisk.R
+import com.topjohnwu.magisk.extensions.rawResource
+import com.topjohnwu.magisk.extensions.toShellCmd
 import com.topjohnwu.superuser.Shell
 import com.topjohnwu.superuser.ShellUtils
 import com.topjohnwu.superuser.io.SuFile
